+++ TestCases/MipsTests/PrintTests/tests/test38.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
 int main(){
-    printf("test: %d %f\n", 1, 0.5, 2);
+    printf("test: %d %d\n", 1, (int) 0.5, 2);
     return 0;
 }
